% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pi.R
\name{extract_median_surv}
\alias{extract_median_surv}
\title{Functions to extract prediction intervals and observed data}
\usage{
extract_median_surv(km.pi, outtype = c("long", "wide"))
}
\arguments{
\item{outtype}{Specifies whether output will be in long or wide format.}
}
\description{
\lifecycle{deprecated}
Please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\details{
\code{\link{extract_median_surv}} extracts prediction intervals of
median survival times and and the corresponding observed values.
}
