% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pi.R
\name{extract_median_surv}
\alias{extract_median_surv}
\title{Functions to extract prediction intervals and observed data}
\usage{
extract_median_surv(km.pi, outtype = c("long", "wide"))
}
\arguments{
\item{km.pi}{A return object from \code{\link[=calc_km_pi]{calc_km_pi()}} function.}

\item{outtype}{Specifies whether output will be in long or wide format.}
}
\description{
\lifecycle{deprecated}
}
\details{
\code{\link[=extract_median_surv]{extract_median_surv()}} was renamed to \code{\link[=extract_medsurv_pi]{extract_medsurv_pi()}} for function name consistency.

\code{\link[=extract_median_surv]{extract_median_surv()}} extracts prediction intervals of
median survival times and and the corresponding observed values.
}
