% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_sim.R
\name{extract_sim}
\alias{extract_sim}
\alias{extract_hr}
\title{Functions to extract raw simulated samples}
\usage{
extract_sim(sim)

extract_hr(hr.pi)
}
\arguments{
\item{sim}{A `survparamsim` class object generated by \code{\link{surv_param_sim}} function.}

\item{hr.pi}{a return object from \code{\link{calc_hr_pi}} function.}
}
\description{
\code{\link{extract_sim}} extract raw survival time & event status for all simulated subjects.
}
\details{
\code{\link{extract_hr}} extract simulated HRs for all repeated simulations
}
