% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hr_pi.R
\name{plot_hr_pi}
\alias{plot_hr_pi}
\title{Plot HR}
\usage{
plot_hr_pi(hr.pi, show.obs = TRUE)
}
\arguments{
\item{hr.pi}{a return object from \code{\link{calc_hr_pi}} function.}

\item{show.obs}{A logical specifying whether to show observed HR on the plot.
This will have no effect if \code{calc.obs} was set to \code{FALSE} in \code{\link{calc_hr_pi}}.}
}
\description{
Plot HR
}
