% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_km_pi.R
\name{plot_km_pi}
\alias{plot_km_pi}
\title{Plot Kaplan-Meier curves with prediction intervals from parametric bootstrap simulation}
\usage{
plot_km_pi(km.pi, show.obs = TRUE, trunc.sim.censor = TRUE)
}
\arguments{
\item{km.pi}{an output from \code{\link{calc_km_pi}} function.}

\item{show.obs}{A logical specifying whether to show observed K-M curve on the plot.
This will have no effect if \code{calc.obs} was set to \code{FALSE} in \code{\link{calc_km_pi}}.}

\item{trunc.sim.censor}{A logical specifying whether to truncate the simulated
curve at the last time of \code{censor.dur} specified in \code{\link{surv_param_sim}}.}
}
\description{
Need to think about how to apply this for subgroups
}
