% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hr_pi.R, R/calc_km_pi.R,
%   R/survparamsim-methods.R
\name{print.survparamsim.hrpi}
\alias{print.survparamsim.hrpi}
\alias{summary.survparamsim.hrpi}
\alias{print.survparamsim.kmpi}
\alias{summary.survparamsim.kmpi}
\alias{survparamsim-methods}
\alias{print.survparamsim}
\title{Methods for S3 objects in the package}
\usage{
\method{print}{survparamsim.hrpi}(x, ...)

\method{summary}{survparamsim.hrpi}(object, ...)

\method{print}{survparamsim.kmpi}(x, ...)

\method{summary}{survparamsim.kmpi}(object, ...)

\method{print}{survparamsim}(x, ...)
}
\arguments{
\item{x}{An object of the corresponding class}

\item{...}{Additional arguments passed to methods.}

\item{object}{An object of the corresponding class}
}
\description{
Methods for S3 objects in the package
}
