% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_sim.R
\name{extract_sim}
\alias{extract_sim}
\alias{extract_hr}
\alias{extract_km_obs}
\title{Functions to extract raw simulated samples}
\usage{
extract_sim(sim)

extract_hr(hr.pi)

extract_km_obs(km.pi)
}
\arguments{
\item{sim}{A \code{survparamsim} class object generated by \code{\link{surv_param_sim}} function.}

\item{hr.pi}{a return object from \code{\link{calc_hr_pi}} function.}

\item{km.pi}{A return object from \code{\link{calc_km_pi}} function.}
}
\description{
Functions to extract raw simulated samples
}
\details{
\code{\link{extract_sim}} extracts raw survival time & event status for all simulated subjects.

\code{\link{extract_hr}} extracts simulated HRs for all repeated simulations

\code{\link{extract_km_obs}} extracts observed Kaplan-Meier curves.
}
